requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
rembg>=2.0.55
pillow>=10.1.0
trimesh>=4.0.5
numpy>=1.26.2
torch>=2.1.1
torchvision>=0.16.1
torchaudio>=2.1.1
einops>=0.7.0
omegaconf>=2.3.0
transformers>=4.36.2
huggingface-hub>=0.20.1
xatlas>=0.0.9
moderngl>=5.8.0
